+++ spec/g5_integrations_validations/multi_family_inventory_rate_to_display_validations_spec.rb
@@ -1,14 +1,14 @@
 require "spec_helper"
 
 module G5IntegrationsValidations
-  describe YardiRentCafeInventoryRateToDisplayValidations do
+  describe MultiFamilyInventoryRateToDisplayValidations do
 
     subject { YardiRentCafeInventory.new }
 
     it do
       is_expected.
         to validate_inclusion_of(:rate_to_display).
-        in_array(YardiRentCafeInventoryRateToDisplayValidations::VALID_OPTIONS)
+        in_array(MultiFamilyInventoryRateToDisplayValidations::VALID_OPTIONS)
     end
 
   end